// src/types/knmi-api.ts
export interface components {
	schemas: {
// ...
		ErrorResponse: {
			/**
			 * Format: int32
			 * @description The HTTP status code
			 */
			status: number;
			/** @description Human-readable error description */
			title: string;
			/** @description Auto-generated ID identifying this request */
			correlationId: string;
		};
		/**
		 * @description The unit of the measurement.
		 *
		 * @enum {string}
		 */
		MeasurementUnit:
			| "beaufort"
			| "kilometerPerHour"
			| "degreeCelsius"
			| "percentage";
// ...
	};
	responses: never;
	parameters: {
		/** @description A unique user identifier, ideally a UUID >= v5 */
		SubscriberId: string;
	};
	requestBodies: never;
	headers: never;
	pathItems: never;
}